requirements.txt
dnslib
PyYAML
